Fixed broken code for importing certificates
The example script is broken and when checking if it a valid certificate, it modifies the certificate by replacing new line '\n' with blank spaces which corrupts the certificate. The check for a valid certificate is meaningless because nothing happens if it is invalid. Therefore, I simply commented out the 2 line that change the certificate and now it imports a certificate properly.